feat: modifying variable(progress) type.

+++ src/AElf.CrossChainServer.Application.Contracts/CrossChain/CrossChainTransferStatusDto.cs
@@ -5,5 +5,5 @@ namespace AElf.CrossChainServer.CrossChain;
 
 public class CrossChainTransferStatusDto : EntityDto<Guid>
 {
-    public double Progress { get; set; }
+    public int Progress { get; set; }
 }
